requirements.txt
flask==3.0.3
google-cloud-bigquery==3.25.0
requests==2.32.3
gunicorn==21.2.0
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
google-auth==2.25.2

